Small fixes

+++ src/Modules/player/AudioResource.ts
@@ -1,6 +1,5 @@
 import {ChildProcessWithoutNullStreams, spawn} from "child_process";
 import {OpusEncoder} from "@watklok/opus";
-import {db} from "@Client/db";
 
 /**
  * @author SNIPPIK
@@ -9,7 +8,7 @@ import {db} from "@Client/db";
  */
 export class AudioResource {
     private readonly _streams = {
-        opus: new OpusEncoder(),
+        opus:    new OpusEncoder(),
         process: null as Process
     };
     private readonly _temp = {
@@ -63,26 +62,24 @@ export class AudioResource {
      * @param options {object} Параметры для создания
      * @public
      */
-    public constructor(options: { path: string; filters?: Filter[]; seek?: number; }) {
+    public constructor(options: { path: string; filters?: string; seek?: number; chunkSize: number; }) {
         const {seek, path} = options;
-        const { filters, speed} = Filters.getParameters(options.filters);
 
-        if (speed > 0) this._temp.frame = 20 * speed;
+        if (options.chunkSize > 0) this._temp.frame = 20 * options.chunkSize;
         if (seek > 0) this._temp.frames = (seek * 1e3) / this._temp.frame;
 
-        this.stream.once("readable", () => {
-            this._temp.readable = true;
-        });
+        //Слушаем OpusEncoder
+        ["end", "close", "error"].forEach((event) => this.stream.once(event, this.cleanup));
+        this.stream.once("readable", () => { this._temp.readable = true; });
 
-        //Запускаем процесс FFmpeg и подключаем его к декодеру
+        //Запускаем процесс FFmpeg и подключаем его к OpusEncoder'у
         const urls = path.split(":|");
         this._streams.process = new Process(["-vn", "-loglevel", "panic",
             ...(urls[0] === "link" ? ["-reconnect", "1", "-reconnect_streamed", "1", "-reconnect_delay_max", "5"] : []),
-            "-ss", `${seek ?? 0}`, "-i", urls[1], "-af", filters, "-f", `${OpusEncoder.lib.ffmpeg}`, "pipe:1"
+            "-ss", `${seek ?? 0}`, "-i", urls[1],
+            ...(options.filters ? ["-af", options.filters]: []),
+            "-f", `${OpusEncoder.lib.ffmpeg}`, "pipe:1"
         ]);
-
-        //Слушаем декодер
-        ["end", "close", "error"].forEach((event) => this.stream.once(event, this.cleanup));
         this._streams.process.stderr.once("error", (err) => this.stream.emit("error", err));
         this._streams.process.stdout.pipe(this.stream);
     };
@@ -94,10 +91,8 @@ export class AudioResource {
     public cleanup = () => {
         for (let item of Object.keys(this._temp)) this._temp[item] = null;
         for (let [key, value] of Object.entries(this._streams)) {
-            if (value instanceof Process) {
-                value.process.emit("close");
-                value.stdout.removeAllListeners();
-            } else {
+            if (value instanceof Process) value.cleanup();
+            else {
                 value?.destroy();
                 value?.removeAllListeners();
             }
@@ -142,80 +137,27 @@ export class Process {
      */
     public get process() { return this._process; };
 
+    /**
+     * @description Задаем параметры и запускаем процесс
+     * @param args {string[]} Аргументы для запуска
+     * @param name {string} Имя процесса
+     */
     public constructor(args: string[], name: string = "ffmpeg") {
         this._process = spawn(name, args);
-        this.process.once("close", this._cleanup);
+
+        ["end", "close", "error"].forEach((event) => this.process.once(event, this.cleanup));
     };
 
     /**
      * @description Удаляем FFmpeg
      * @private
      */
-    private _cleanup = () => {
-        if (!this._process?.killed) {
-            this.stdout.destroy();
-            this.stderr.destroy();
-            this.stdin.destroy();
-
-            this._process.kill();
+    public cleanup = () => {
+        for (const std of [this.stdout, this.stdin, this.stderr]) {
+            std.removeAllListeners();
+            std.destroy();
         }
+
+        if (!this._process?.killed) this._process.kill();
     };
-}
-
-/**
- * @author SNIPPIK
- * @class Filters
- */
-export const Filters = new class {
-    /**
-     * @description Получаем параметры фильтров ит готовые фильтры для FFmpeg
-     * @param filters {Filter[]} Фильтры
-     * @return object
-     */
-    public getParameters = (filters: Filter[]) => {
-        const realFilters = [`volume=${db.AudioOptions.volume / 100}`]; let speed = 0;
-
-        //Проверяем фильтры
-        for (const filter of filters) {
-
-            //Если фильтр не требует аргумента
-            if (!filter.args) realFilters.push(filter.filter);
-            else realFilters.push(filter.filter + filter.user_arg ?? "");
-
-            //Если у фильтра есть модификатор скорости
-            if (filter?.speed) {
-                if (typeof filter.speed === "number") speed += Number(filter.speed);
-                else speed += Number(filters.slice(filters.indexOf(filter) + 1));
-            }
-        }
-
-        //Надо ли плавное включения треков
-        realFilters.push(`afade=t=in:st=0:d=${db.AudioOptions.fade}`);
-
-        return { filters: realFilters.join(","), speed }
-    }
-}
-
-/**
- * @author SNIPPIK
- * @description Как выглядит фильтр
- * @interface
- */
-export interface Filter {
-    //Имена
-    names: string[];
-
-    //Описание
-    description: string;
-
-    //Сам фильтр
-    filter: string;
-
-    //Аргументы
-    args: false | [number, number];
-
-    user_arg: any;
-
-    //Меняется ли скорость
-    speed?: null | number;
 }
